BAdapter

Description

Exports the IsAdapter class. Defines a functions an Adapter should have by default.

Documentation

class IsAdapter a where

Any Adapter should have these methods... _init: Initialize the adatper hear: Listen for events and inputs

Minimal complete definition

_init, hear

Methods

_init :: a -> IO ()

_init: Initialize an Adapter. Display a start message, etc.

hear :: a -> IO Text

hear: receive an Adapeter as input, along with the State, and do IO etc